+++ test/maximum_jank_test.exs
@@ -1,6 +1,14 @@
 defmodule MaximumJankTest do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
   import MaximumJank
+
+  describe "&module!/1" do
+    test "returns module",  do: assert module!(MaximumJank) == MaximumJank
+
+    test "raises an error" do
+      assert_raise RuntimeError, ~r/not a module/, fn -> module!(FakeModule) end
+    end
+  end
 
   describe "&module?/1" do
     test "is true",  do: assert module?(MaximumJank) == true