+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
-## 0.1.0 / 2017-08-30
+## 0.1.0 / 2017-09-03 
 ### Added
-- &MaximumJank.module?/1
+- `&MaximumJank.module!/1`
+- `&MaximumJank.module?/1`